Search keywords on paste
