+++ src/background/usecases/parsers.js
@@ -1,27 +1,3 @@
-const trimStart = (str) => {
-  // NOTE String.trimStart is available on Firefox 61
-  return str.replace(/^\s+/, '');
-};
-
-const normalizeUrl = (keywords, searchSettings) => {
-  try {
-    return new URL(keywords).href;
-  } catch (e) {
-    if (keywords.includes('.') && !keywords.includes(' ')) {
-      return 'http://' + keywords;
-    }
-    let template = searchSettings.engines[searchSettings.default];
-    let query = keywords;
-
-    let first = trimStart(keywords).split(' ')[0];
-    if (Object.keys(searchSettings.engines).includes(first)) {
-      template = searchSettings.engines[first];
-      query = trimStart(trimStart(keywords).slice(first.length));
-    }
-    return template.replace('{}', encodeURIComponent(query));
-  }
-};
-
 const mustNumber = (v) => {
   let num = Number(v);
   if (isNaN(num)) {
@@ -52,4 +28,4 @@ const parseSetOption = (word, types) => {
   }
 };
 
-export { normalizeUrl, parseSetOption };
+export { parseSetOption };
